requirements.txt
pyspark==3.5.0
delta-spark==3.1.0
numpy>=2.0.0
pandas>=2.3.0
scipy>=1.13.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
notebook>=7.0.0
requests>=2.31.0
